Lift ImportDepth utility out of SourceKit into IDE (NFC)
