+++ include/swift/IDE/CodeCompletion.h
@@ -17,6 +17,7 @@
 #include "swift/Basic/Debug.h"
 #include "swift/Basic/LLVM.h"
 #include "llvm/ADT/ArrayRef.h"
+#include "llvm/ADT/StringMap.h"
 #include "llvm/ADT/StringRef.h"
 #include "llvm/Support/Allocator.h"
 #include "llvm/Support/TrailingObjects.h"
@@ -29,6 +30,7 @@ namespace swift {
 class CodeCompletionCallbacksFactory;
 class Decl;
 class DeclContext;
+class FrontendOptions;
 class ModuleDecl;
 
 namespace ide {
@@ -863,6 +865,26 @@ struct CodeCompletionResultSink {
 
   CodeCompletionResultSink()
       : Allocator(std::make_shared<llvm::BumpPtrAllocator>()) {}
+};
+
+/// A utility for calculating the import depth of a given module. Direct imports
+/// have depth 1, imports of those modules have depth 2, etc.
+///
+/// Special modules such as Playground auxiliary sources are considered depth
+/// 0.
+class ImportDepth {
+  llvm::StringMap<uint8_t> depths;
+
+public:
+  ImportDepth() = default;
+  ImportDepth(ASTContext &context, const FrontendOptions &frontendOptions);
+
+  Optional<uint8_t> lookup(StringRef module) {
+    auto I = depths.find(module);
+    if (I == depths.end())
+      return None;
+    return I->getValue();
+  }
 };
 
 class CodeCompletionContext {
