+++ tools/SourceKit/lib/SwiftLang/CodeCompletionOrganizer.cpp
@@ -55,20 +55,6 @@ struct Result : public Item {
     return item->getKind() == ItemKind::Result;
   }
 };
-class ImportDepth {
-  llvm::StringMap<uint8_t> depths;
-
-public:
-  ImportDepth() = default;
-  ImportDepth(ASTContext &context, const CompilerInvocation &invocation);
-
-  Optional<uint8_t> lookup(StringRef module) {
-    auto I = depths.find(module);
-    if (I == depths.end())
-      return None;
-    return I->getValue();
-  }
-};
 } // end anonymous namespace
 
 struct CodeCompletion::Group : public Item {
@@ -93,7 +79,8 @@ std::vector<Completion *> SourceKit::CodeCompletion::extendCompletions(
   ImportDepth depth;
   if (info.swiftASTContext) {
     // Build import depth map.
-    depth = ImportDepth(*info.swiftASTContext, *info.invocation);
+    depth = ImportDepth(*info.swiftASTContext,
+                        info.invocation->getFrontendOptions());
   }
 
   if (info.completionContext)
@@ -328,75 +315,6 @@ void CodeCompletionOrganizer::groupAndSort(const Options &options) {
 
 CodeCompletionViewRef CodeCompletionOrganizer::takeResultsView() {
   return impl.takeView();
-}
-
-//===----------------------------------------------------------------------===//
-// ImportDepth
-//===----------------------------------------------------------------------===//
-
-ImportDepth::ImportDepth(ASTContext &context,
-                         const CompilerInvocation &invocation) {
-  llvm::DenseSet<ModuleDecl *> seen;
-  std::deque<std::pair<ModuleDecl *, uint8_t>> worklist;
-
-  StringRef mainModule = invocation.getModuleName();
-  auto *main = context.getLoadedModule(context.getIdentifier(mainModule));
-  assert(main && "missing main module");
-  worklist.emplace_back(main, uint8_t(0));
-
-  // Imports from -import-name such as Playground auxiliary sources are treated
-  // specially by applying import depth 0.
-  llvm::StringSet<> auxImports;
-  for (const auto &pair :
-       invocation.getFrontendOptions().getImplicitImportModuleNames())
-    auxImports.insert(pair.first);
-
-  // Private imports from this module.
-  // FIXME: only the private imports from the current source file.
-  // FIXME: ImportFilterKind::ShadowedByCrossImportOverlay?
-  SmallVector<ImportedModule, 16> mainImports;
-  main->getImportedModules(mainImports,
-                           {ModuleDecl::ImportFilterKind::Default,
-                            ModuleDecl::ImportFilterKind::ImplementationOnly});
-  for (auto &import : mainImports) {
-    uint8_t depth = 1;
-    if (auxImports.count(import.importedModule->getName().str()))
-      depth = 0;
-    worklist.emplace_back(import.importedModule, depth);
-  }
-
-  // Fill depths with BFS over module imports.
-  while (!worklist.empty()) {
-    ModuleDecl *module;
-    uint8_t depth;
-    std::tie(module, depth) = worklist.front();
-    worklist.pop_front();
-
-    if (!seen.insert(module).second)
-      continue;
-
-    // Insert new module:depth mapping.
-    const clang::Module *CM = module->findUnderlyingClangModule();
-    if (CM) {
-      depths[CM->getFullModuleName()] = depth;
-    } else {
-      depths[module->getName().str()] = depth;
-    }
-
-    // Add imports to the worklist.
-    SmallVector<ImportedModule, 16> imports;
-    module->getImportedModules(imports);
-    for (auto &import : imports) {
-      uint8_t next = std::max(depth, uint8_t(depth + 1)); // unsigned wrap
-
-      // Implicitly imported sub-modules get the same depth as their parent.
-      if (const clang::Module *CMI =
-              import.importedModule->findUnderlyingClangModule())
-        if (CM && CMI->isSubModuleOf(CM))
-          next = depth;
-      worklist.emplace_back(import.importedModule, next);
-    }
-  }
 }
 
 //===----------------------------------------------------------------------===//
